refactor(pull-request): move pr comment to separate file
